Update styles, prop types and tests
