--- src/components/MobilityPlatform/ParkingChargeZones/components/ParkingChargeZoneContent/styles.js
@@ -1,11 +1,0 @@
-export default theme => ({
-  padding: {
-    padding: theme.spacing(1),
-  },
-  subtitle: {
-    marginBottom: theme.spacing(1),
-    paddingBottom: theme.spacing(0.5),
-    borderBottom: '1px solid #000000',
-    width: '88%',
-  },
-});
