+++ src/views/MobilitySettingsView/components/ParkingChargeZoneList/index.js
@@ -1,6 +1,3 @@
-import { withStyles } from '@mui/styles';
-import { injectIntl } from 'react-intl';
 import ParkingChargeZoneList from './ParkingChargeZoneList';
-import styles from './styles';
 
-export default withStyles(styles)(injectIntl(ParkingChargeZoneList));
+export default ParkingChargeZoneList;
