--- src/views/MobilitySettingsView/components/ParkingChargeZoneList/styles.js
@@ -1,11 +1,0 @@
-const styles = theme => ({
-  checkBoxContainer: {
-    borderBottom: '1px solid #6f7276',
-    display: 'flex',
-    flexDirection: 'column',
-    justifyContent: 'start',
-    paddingLeft: theme.spacing(3.5),
-  },
-});
-
-export default styles;
